add bare minimum to make TUI work
By following the tutorial from
https://blog.logrocket.com/rust-and-tui-building-a-command-line-interface-in-rust/
we have been able to make a functional TUI. There is a lot of work to
do to improve this but this should work to understand the basics.

Also, there we are adding some domain ideas to start to model the Cell
structs. This is still WIP.